+++ src/App.tsx
@@ -1,4 +1,4 @@
-import Main from './Main';
+import Main from "./Main.tsx";
 
 function App() {
   return (
