+++ src/constants.ts
@@ -1,3 +1,3 @@
-export const API_BASE_URL = 'https://bootcamp-api.codeit.kr/api'
+export const API_BASE_URL = "https://bootcamp-api.codeit.kr/api";
 
-export const ASSETS_URL= process.env.PUBLIC_URL;
+export const ASSETS_URL = process.env.PUBLIC_URL;
