+++ src/index.tsx
@@ -1,11 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-import App from './App';
-import './index.css';
-import CreateDOM from 'react-dom/client';
+import React from "react";
+import { BrowserRouter } from "react-router-dom";
+import CreateDOM from "react-dom/client";
+import App from "./App.tsx";
+import "./index.css";
 
-const root = CreateDOM.createRoot(document.getElementById('root') as HTMLElement);
+const root = CreateDOM.createRoot(
+  document.getElementById("root") as HTMLElement
+);
 
 root.render(
   <React.StrictMode>
