+++ src/services/GetData.tsx
@@ -1,11 +1,14 @@
-import { API_BASE_URL } from '../constants';
+import { API_BASE_URL } from "../constants.ts";
 
-export async function getDate(url: string) {
+async function getDate(url: string) {
   try {
     const response = await fetch(`${API_BASE_URL}${url}`);
     if (!response.ok) throw new Error();
     return await response.json();
   } catch (error) {
-    console.log(error);
+    console.log("An error occurred:", error);
+    throw error;
   }
 }
+
+export default getDate;
